% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{baseXmlToList}
\alias{baseXmlToList}
\title{Reads base xml file and creates list to be edited with
openMalariaUtilities}
\usage{
baseXmlToList(filename)
}
\arguments{
\item{filename}{Name of OpenMalaria base xml file}
}
\description{
Reads base xml file and creates list to be edited with
openMalariaUtilities
}
