% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing_aggrFun.R
\name{CalcEpiOutputs}
\alias{CalcEpiOutputs}
\title{Calculate epidemiological indicators from stacked raw data}
\usage{
CalcEpiOutputs(
  df,
  indicators = c("incidenceRate", "prevalenceRate"),
  aggregateByAgeGroup = NULL,
  aggregateByDate = NULL,
  timeHorizon = NULL,
  use.gc = FALSE
)
}
\arguments{
\item{df}{Input data frame}

\item{indicators}{Can be a character vector of epidemiological indicators to
calculate or measures from OpenMalaris. The calculated indicators can be
any of:
prevalenceRate
incidenceRate
incidenceRatePerThousand
tUncomp
tSevere
nHosp
edeath
edeathRatePerHundredThousand
edirdeath
edirdeathRatePerHundredThousand
ddeath
ddeathRatePerHundredThousand}

\item{aggregateByAgeGroup}{Character vector of age groups (e.g. "2-10" for
the right-open interval) to be used for aggregation, if NULL, then age
groups from surveys will be used}

\item{aggregateByDate}{If NULL, keep original surveys. Otherwise any
combination of:

\if{html}{\out{<div class="sourceCode">}}\preformatted{   - "year" aggregate by year (date will be changed to 31-12 of the
     corresponding year).

   - "month" aggregate by month (date will be changed to the 15th of a
     month).

   - "survey" keep survey dates.
}\if{html}{\out{</div>}}}

\item{timeHorizon}{Character vector of with start date and end date of time
horizon to be processed, e.g. c("2000-01-01", "2100-01-01")}

\item{use.gc}{If TRUE, run garbage collection during operation to keep
memory consumption low. Useful if computational ressources are tight.}
}
\description{
Calculate epidemiological indicators from stacked raw data
}
