% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing-db.R
\name{calculateEpidemiologicalIndicators_old}
\alias{calculateEpidemiologicalIndicators_old}
\title{Calculate epidemiological indicators from stacked raw data}
\usage{
calculateEpidemiologicalIndicators_old(
  dbName = NULL,
  experiment_id = NULL,
  scenario_id = NULL,
  metadataFeatures = NULL,
  indicators = c("incidence", "prevalence", "simulatedEIR"),
  customIndicatorFormula = NULL,
  keepMeasuresForIndicators = NULL,
  aggregateByAgeGroup = NULL,
  aggregateByDate = NULL,
  wideFormat = TRUE,
  timeHorizon = c("2000-01-01", "2100-01-01")
)
}
\arguments{
\item{dbName}{Name of sqlite data base file of output results, if null
search of sqlite file in rootDir}

\item{experiment_id}{Number of experiment id to process}

\item{scenario_id}{Numeric vector of scenario ids to process, if NULL then process all scenarios within given experiment}

\item{metadataFeatures}{Character vector of metadata features to keep, if null keep all}

\item{indicators}{Character vector of epidemiological indicators to calculate}

\item{customIndicatorFormula}{Character vector to parse and evaluate for custom indicators}

\item{keepMeasuresForIndicators}{List of indicators whose measures you want to keep in the function output, if null keep none}

\item{aggregateByAgeGroup}{Character vector of age groups (e.g. "2-10" for the
right-open interval) to be
used for aggregation, if NULL, then age groups from surveys will be used}

\item{aggregateByDate}{List (named "month", "year", "date") list(month="month",year="year",date=as.Date(c("2010-04-06","2012-04-06"))) to
be used for aggregation, if NULL, then survey dates will be used, for month, aggregated survey date is set to middle of month}

\item{wideFormat}{Logical vector, if TRUE than output in wide format}

\item{timeHorizon}{Character vector of with start date and end date of
time horizon to be processed, by default start on 2000-01-01}
}
\description{
Calculate epidemiological indicators from stacked raw data
}
