% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_new.R
\name{calibrate_country}
\alias{calibrate_country}
\title{Calibrate a country}
\usage{
calibrate_country(
  simul_df,
  data_to_fit,
  calibration_spec = list(cols_observed = "PR_obs", cols_simulated = "PR"),
  confidence = 0.95,
  lambda = 1,
  select_subset = 0.5,
  llk_type = "gaussian",
  min_EIRnb_to_keep = 10,
  min_possible_EIR = 1,
  max_possible_EIR,
  lower_threshold = 3,
  lower_threshold2 = NULL,
  upper_threshold,
  upper_threshold2 = NULL
)
}
\arguments{
\item{simul_df}{A dataframe containing all the OpenMalaria simulations
(output of get_all_OMsimul)}

\item{data_to_fit}{The dataset to which the simulations are fitted}

\item{calibration_spec}{A list specifying the column names of the the
observations and the simulations, e.g. list(cols_observed = "PR_obs",
cols_simulated = "PR")}

\item{confidence}{Level of statistical confidence for the interval, default
is 95\%}

\item{lambda}{Proportion of the sample to be kept during the smoothing phase
(default is 1)}

\item{select_subset}{Criterion to select only a subset of the EIR values
which are near the maximum, and to select a larger sample for large EIRs
and a smaller sample for small EIRs. To be adjusted by the user to get an
acceptable quadratic approximation}

\item{llk_type}{Choose the likelihood to be used ("gaussian" or "laplace")}

\item{min_EIRnb_to_keep}{Minimum number of EIR values to keep, regardless of
the chosen window (default is 10 and should not need to be adjusted in
practice)}

\item{min_possible_EIR}{Lowest EIR value in the simulated grid (used when
the lower bound of the confidence interval is either negative or the best
EIR is below lower_threshold)}

\item{max_possible_EIR}{Lowest EIR value in the simulated grid (used when
the best EIR is above upper_threshold)}

\item{lower_threshold}{Threshold for the best EIR value under which the
min_possible_EIR is used as lower bound of the confidence interval}

\item{lower_threshold2}{Threshold for the best EIR value above which the
min_possible_EIR is used as lower bound of the confidence interval, in
case the lower bound is NA}

\item{upper_threshold}{Threshold for the best EIR value above which the
max_possible_EIR is used as upper bound of the confidence interval}

\item{upper_threshold2}{Threshold for the best EIR value above which the
max_possible_EIR is used as upper bound of the confidence interval, in
case the upper bound is NA}
}
\description{
This function is based on the methodology by Ionides et al.
2017 (https://doi.org/10.1098/rsif.2017.0126) to calculate confidence
intervals based on the approximate profile likelihood, accounting for
Monte Carlo error. It includes a first smoothing step of the likelihood
function, using the loess function. The calculation of the confidence
intervals relies on a local quadratic approximation, where local refers to
a selection window that can be tuned by the user with the parameters
select_subset and lambda.
}
\section{Required column names}{
Currently, 'calibrate_country()' expects the following columns names in
the input data sets:
simul_df: setting, year, sub, age, PR
data_to_fit: setting, year, sub, age, PR_obs, sd
}

