% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{.defineInterventionsHeader}
\alias{.defineInterventionsHeader}
\title{Make sure required elements are declared}
\usage{
.defineInterventionsHeader(baseList, name = "All interventions")
}
\arguments{
\item{baseList}{List with experiment data.}

\item{name}{Name of the interventions.}
}
\description{
Make sure required elements are declared
}
\keyword{internal}
