% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{get_all_OMsimul}
\alias{get_all_OMsimul}
\title{Retrieve postprocessed results and prepare for calibration}
\usage{
get_all_OMsimul(
  experimentName = NULL,
  workingDir = NULL,
  experimentDir = NULL,
  EIR_name = "scaledAnnualEIR"
)
}
\arguments{
\item{workingDir}{working directory, in null then current working directory}

\item{experimentDir}{full path of experiment directory}

\item{EIR_name}{the name you gave to scaledAnnualEIR in your scenarios table}

\item{experimenName}{experiment name relative to working directory, if null then obtain from experimentDir}
}
\value{
Dataframe
}
\description{
Retrieve postprocessed results and prepare for calibration
}
