% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{get_best_EIR}
\alias{get_best_EIR}
\title{Calculate the likelihood of the simulation with the data, for all EIR}
\usage{
get_best_EIR(simul_df, data_to_fit = fitdat)
}
\arguments{
\item{simul_df}{A dataframe containing all the OpenMalaria simulations
(output of get_all_OMsimul)}

\item{data_to_fit}{The dataset to which the simulations are fitted}
}
\value{
A list of 3 dataframes:
\enumerate{
\item Best_EIR_ML_gaussian" has the simulations with the best EIR according
to the gaussian likelihood best_EIR_ML_laplace"
\item Has the simulations with the best EIR according to the laplace
likelihood
\item All_llk includes all the simulations and their associated likelihood
values
}
}
\description{
Calculate the likelihood of the simulation with the data, for all EIR
}
