% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{get_profile_llk_ci_approx}
\alias{get_profile_llk_ci_approx}
\title{Selection of the upper and lower bound on the EIR and merge with the
simulation file}
\usage{
get_profile_llk_ci_approx(
  my_quadratic_approx,
  min_possible_EIR = 1,
  max_possible_EIR = 200,
  lower_threshold = 3,
  upper_threshold = 190,
  upper_threshold2 = 160
)
}
\arguments{
\item{my_quadratic_approx}{Output from the function get_quadratic_fit_cf}

\item{min_possible_EIR}{Lowest EIR value in the simulated grid (used when the
lower bound of the confidence interval is either negative or the best EIR
is below lower_threshold)}

\item{max_possible_EIR}{Lowest EIR value in the simulated grid (used when the
best EIR is above upper_threshold)}

\item{lower_threshold}{Threshold for the best EIR value under which the
min_possible_EIR is used as lower bound of the confidence interval}

\item{upper_threshold}{Threshold for the best EIR value above which the
max_possible_EIR is used as upper bound of the confidence interval}

\item{upper_threshold2}{Threshold for the best EIR value above which the
max_possible_EIR is used as upper bound of the confidence interval, in case
the upper bound is NA}
}
\value{
A dataframe with one line per subsetting, indicating the best EIR
(EIRmle_tilde, alonside its rounded value EIRmle_tilde_round), the upper
bound (EIR_uci) and the lower bound (EIR_lci) of the confidence interval.
It also contains the associated likelihood values.
}
\description{
Post-processes the output of get_quadratic_fit_cf to retrieve the
best EIR and its confidence interval. Additional parameters are required to
specifiy the rules to be applied at the edges of the EIR grid
}
