% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{get_quadratic_fit_cf}
\alias{get_quadratic_fit_cf}
\title{Get the quadratic fit, following the methodology by Ionides et al.
2017}
\usage{
get_quadratic_fit_cf(
  df,
  df_all,
  confidence = 0.95,
  lambda = 1,
  select_subset = 0.5,
  llk_type = "llk_gaussian",
  data_to_fit = fitdat,
  min_EIRnb_to_keep = 10
)
}
\arguments{
\item{df}{Best EIR file, as output of get_best_EIR (could be the
best_EIR_ML_gaussian or the best_EIR_ML_gaussian output)}

\item{df_all}{The all_llk output of the get_best_EIR function}

\item{confidence}{Level of statistical confidence for the interval, default
is 95\%}

\item{lambda}{Proportion of the sample to be kept during the smoothing phase
(default is 1)}

\item{select_subset}{Criterion to select only a subset of the EIR values
which are near the maximum, and to select a larger sample for large EIRs
and a smaller sample for small EIRs. To be adjusted by the user to get an
acceptable quadratic approximation}

\item{llk_type}{Choose the likelihood to be used (default "llk_gaussian" is
gaussian)}

\item{data_to_fit}{The dataset to which the simulations are fitted}

\item{min_EIRnb_to_keep}{Minimum number of EIR values to keep, regardless of
the chosen window (default is 10 and should not need to be adjusted in
practice)}
}
\value{
a dataframe with the fitted quadratic approximation. This output is
needed as input for the function get_profile_llk_ci_approx and is needed to
plot diagnostics on the quality of the quadratic approximation.
}
\description{
This function is based on the methodology by Ionides et al. 2017
(https://doi.org/10.1098/rsif.2017.0126) to calculate confidence intervals
based on the approximate profile likelihood, accounting for Monte Carlo
error. It includes a first smoothing step of the likelihood function, using
the loess function. The calculation of the confidence intervals relies on a
local quadratic approximation, where local refers to a selection window
that can be tuned by the user with the parameters select_subset and lambda.
}
