% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing-db.R
\name{measuresNeededForPostprocessing_old}
\alias{measuresNeededForPostprocessing_old}
\title{List of measures needed to calculate epidemiological indicators}
\usage{
measuresNeededForPostprocessing_old()
}
\description{
List of measures needed to calculate epidemiological indicators
}
\keyword{internal}
