% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EntomologyVectorControl.R
\name{mosquitoParameterization}
\alias{mosquitoParameterization}
\title{Creates MosquitoParameters list from AnophelesModel package
used to define mosquito bionomics with defineEntomology function}
\usage{
mosquitoParameterization(vectorSpecies, OM_snippet = T, verbatim = F)
}
\arguments{
\item{vectorSpecies}{vector species to be considered}

\item{OM_snippet}{If TRUE, then output is snippet for OpenMalaria}

\item{verbatim}{If TRUE, then show messages}
}
\description{
Creates MosquitoParameters list from AnophelesModel package
used to define mosquito bionomics with defineEntomology function
}
\details{
vectorSpecies=c("Anopheles gambiae","Anopheles funestus"
}
