% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm_postprocessing.R
\name{slurmPreparePostprocessing}
\alias{slurmPreparePostprocessing}
\title{Run preparations for SLURM submission for postprocessing}
\usage{
slurmPreparePostprocessing(
  expName,
  dbName,
  bSize = 200,
  ntasks = 1,
  memCPU = "32G",
  nCPU = 1,
  time = "06:00:00",
  qos = "6hours",
  readResults = FALSE,
  experiment_id = 1,
  scenario_id = NULL,
  indicators = c("simulatedEIR", "incidence", "prevalence"),
  customIndicatorFormula = list(incidencePerFifty = "(nUncomp+nSevere)/nHost*50"),
  metadataFeatures = c("setting", "EIRid"),
  keepMeasuresForIndicators = NULL,
  aggregateByAgeGroup = c("0-5", "2-10"),
  aggregateByDate = list(month = "month", year = "year", date = as.Date(c("2010-04-06",
    "2012-04-06", "2015-01-06", "2030-04-06"))),
  wideFormat = TRUE,
  timeHorizon = c("2000-01-01", "2100-01-01")
)
}
\arguments{
\item{expName}{Name of experiment}

\item{dbName}{Name of sqlite data base file of output results, if null
search of sqlite file in rootDir}

\item{bSize}{Number of parallely processed scenario outputs, if null defaults
to batching used for scenarios}

\item{ntasks}{Number of tasks per CPU}

\item{memCPU}{Memory per CPU}

\item{nCPU}{Number of cores to reserve}

\item{time}{Maximum time}

\item{qos}{Quality of service}

\item{readResults}{Extract data and overwrites sqlite file while creating
data base from output files, same as slurmPrepareResults, slurmRunResults}

\item{experiment_id}{Numeric vector of experiment ids to process}

\item{scenario_id}{Numeric vector of scenario ids to process}

\item{indicators}{Character vector of epidemiological indicators to calculate}

\item{customIndicatorFormula}{Character vector to parse and evaluate for
custom indicators}

\item{metadataFeatures}{Character vector of metadata features (patterns are
matched) to keep, if null keep all}

\item{keepMeasuresForIndicators}{Character vector of indicators whose
measures we want to keep}

\item{aggregateByAgeGroup}{Character vector of age groups (e.g. "2-10" for
the right-open interval) to be
used for aggregation, if NULL, then age groups from surveys will be used}

\item{aggregateByDate}{List (named "month", "year", "date") list(month="month",year="year",date=as.Date(c("2010-04-06","2012-04-06"))) to
be used for aggregation, if NULL, then survey dates will be used, for month, aggregated survey date is set to middle of month}

\item{wideFormat}{Logical vector, if TRUE than output in wide format}

\item{timeHorizon}{Character vector of start date and end date of
time horizon to be processed, by default start on 2000-01-01}
}
\description{
Run preparations for SLURM submission for postprocessing
}
