% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm_postprocessing.R
\name{slurmRunPostprocessing}
\alias{slurmRunPostprocessing}
\title{Submit postprocessing job to SLURM}
\usage{
slurmRunPostprocessing(wait = FALSE)
}
\arguments{
\item{wait}{If TRUE, use the '--wait' flag}
}
\description{
Submit postprocessing job to SLURM
}
