% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EntomologyVectorControl.R
\name{vectorControlParameterization}
\alias{vectorControlParameterization}
\title{Creates VectorInterventionParameters list from AnophelesModel package
used to define vector control interventions with defineVectorControl function}
\usage{
vectorControlParameterization(
  vectorSpecies,
  interventionList,
  mosq = NULL,
  verbatim = F
)
}
\arguments{
\item{vectorSpecies}{vector species to be considered}

\item{interventionList}{list of interventions to be considered}

\item{mosq}{Existing mosquito parameterization, if NULL then build from
vectorSpecies argument}

\item{verbatim}{If TRUE, then show messages}
}
\description{
Creates VectorInterventionParameters list from AnophelesModel package
used to define vector control interventions with defineVectorControl function
}
\examples{
vectorSpecies=c("Anopheles gambiae","Anopheles funestus")
interventionList=list(LLIN=list(id="LLIN",
description="test LLIN",
parameterisation="LLINs01",
LLIN_type="Default",
LLIN_insecticide="Default",
LLIN_country="Kenya"),
IRS=list(id="IRS",
         description="IRS",
         parameterisation="IRS02")
)
}
